+++ Assets/Source/Logic/Fight/HealthSystems/HurtBoxes/BaseHurtBox.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 [RequireComponent(typeof(Collider))]
 public class BaseHurtBox : MonoCache, ITakeHit
 {
-    public event Action<int> OnTakeHit;
+    public event Action<float> OnTakeHit;
     public event Action<Vector3> OnTakeHitWithPosition;
     
-    public void TakeHit(int damage, Vector3 hitPosition)
+    public void TakeHit(float damage, Vector3 hitPosition)
     {
         OnTakeHit?.Invoke(damage);
         OnTakeHitWithPosition?.Invoke(hitPosition);
