+++ Assets/Source/Logic/Fight/HealthSystems/HurtBoxes/BaseWeakPointHurtBox.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 [RequireComponent(typeof(Collider))]
 public class BaseWeakPointHurtBox : MonoCache, IWeakPoint
 {
-    public event Action<int> OnWeakPointHit;
+    public event Action<float> OnWeakPointHit;
     public event Action<Vector3> OnWeakPointHitWithPosition;
     
-    public void TakeWeakPointHit(int damage, Vector3 hitPosition)
+    public void TakeWeakPointHit(float damage, Vector3 hitPosition)
     {
         OnWeakPointHit?.Invoke(damage);
         OnWeakPointHitWithPosition?.Invoke(hitPosition);
