+++ Assets/Source/Logic/Interfaces/Fight I/IHealthHandler.cs
@@ -5,13 +5,13 @@ public interface IHealthHandler
     public event Action OnDying;
     public event Action OnDied;
 
-    public void HandleHit(int damage);
-    public void HandleWeakPointHit(int baseDamage);
+    public void HandleHit(float damage);
+    public void HandleWeakPointHit(float baseDamage);
     public void StartDying();
     public void Die();
 
-    public void AddHealth(int addValue);
-    public void RemoveHealth(int removeValue);
-    public void SetHealth(int value);
-    public int GetHealth();
+    public void AddHealth(float addValue);
+    public void RemoveHealth(float removeValue);
+    public void SetHealth(float value);
+    public float GetHealth();
 }
