+++ Assets/Source/Logic/Interfaces/Fight I/ITakeHit.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 
 public interface ITakeHit
 {
-    public event Action<int> OnTakeHit;
+    public event Action<float> OnTakeHit;
     public event Action<Vector3> OnTakeHitWithPosition;
 
-    public void TakeHit(int damage, Vector3 hitPosition);
+    public void TakeHit(float damage, Vector3 hitPosition);
 }
