+++ Assets/Source/Logic/Interfaces/Fight I/IWeakPoint.cs
@@ -3,8 +3,8 @@ using UnityEngine;
 
 public interface IWeakPoint
 {
-    public event Action<int> OnWeakPointHit;
+    public event Action<float> OnWeakPointHit;
     public event Action<Vector3> OnWeakPointHitWithPosition;
 
-    public void TakeWeakPointHit(int damage, Vector3 hitPosition);
+    public void TakeWeakPointHit(float damage, Vector3 hitPosition);
 }
